Role enum changed to String and recreated DB
Role was changed on Usuario from Ordinary to String to be able to read it as text.

Database had to be created again by changing the application.properties file and then back it up to validate.

Minor try made on Configuracion to allow the created users in db to be use as login.