Views, controllers and more.
Delete on cascade when deleting a show will delete also the ratings associated to that show.

Views fixed and organized with text centered

Passwords now must match before saving in the database as well as the email must have an email format.

Edit ratings is possible.

And I think that's it. Happy to have worked together with Valentina. What a great IT girl.